# api/openapi.yaml
openapi: 3.1.0
info:
  title: NovaForge API
  version: 0.1.0
  description: >-
    The NovaForge control plane. The web GUI is deferred, so this contract is
    the complete description of the platform for every client: the nf CLI,
    external agents, and the future GUI.
servers:
  - url: /
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: No usable credentials were presented
    Forbidden:
      description: The caller's organization scope does not permit this
    NotFound:
      description: No such resource in the caller's organization
paths:
...
  /api/v1/orgs/{org}/agents:
    get:
      operationId: listAgents
      summary: Agents in an organization
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createAgent
      summary: Define an agent
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/dashboard:
    get:
      operationId: getDashboard
      summary: What needs human attention
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}:
    delete:
      operationId: deleteRepo
      summary: Delete a repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      operationId: getRepo
      summary: One repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/branches:
    get:
      operationId: listBranches
      summary: Branches
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createBranch
      summary: Create a branch
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/maintenance:
    get:
      operationId: listMaintenanceProposals
      summary: What the maintenance scanners proposed
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/maintenance/scan:
    post:
      operationId: scanRepository
      summary: Run the maintenance scanners against a repository now
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/maintenance/{fingerprint}/approve:
    post:
      operationId: approveMaintenanceProposal
      summary: Approve a maintenance proposal and assign its Work Item
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/maintenance/{fingerprint}/dismiss:
    post:
      operationId: dismissMaintenanceProposal
      summary: Dismiss a maintenance proposal, with a reason
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/runs:
    get:
      operationId: listRuns
      summary: Engineering Runs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createRun
      summary: Open an Engineering Run
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/runs/{number}/impact:
    get:
      operationId: getRunImpact
      summary: A run's change impact, measured from its merge base, with a stated risk
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: number
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/runs/{number}/merge:
    post:
      operationId: mergeRun
      summary: Merge, if the gates allow it
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: number
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/runs/{number}/plan:
    get:
      operationId: getRunPlan
      summary: A run's plan
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: number
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/runs/{number}/proof:
    get:
      operationId: getRunProof
      summary: Per-gate proof for a run
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: number
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/runs/{number}/reviews:
    post:
      operationId: submitReview
      summary: Submit a review verdict
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: number
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/work:
    get:
      operationId: listWorkItems
      summary: Work Items
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createWorkItem
      summary: Create a Work Item
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...